display integrated image

+++ notebooks/__code/panoramic_stitching_for_tof/load_data.py
@@ -80,4 +80,4 @@ class LoadData:
 
     @staticmethod
     def calculate_integrated_data(list_arrays=None):
-        return np.sum(list_arrays, 0)
+        return np.sum(list_arrays, 0, dtype=np.int16)
